initial version

+++ src/tk/treectrl/tkTreeCtrl.c
@@ -18,7 +18,6 @@
  * RCS: @(#) $Id: tkTreeCtrl.c,v 1.110 2009/04/18 18:41:52 treectrl Exp $
  */
 
-#include <stdint.h>
 #include "tkTreeCtrl.h"
 
 #ifdef WIN32
@@ -1975,7 +1974,7 @@ Tree_AddItem(
 	int id, isNew;
 
 	id = TreeItem_SetID(tree, item, tree->nextItemId++);
-	hPtr = Tcl_CreateHashEntry(&tree->itemHash, (char *)(intptr_t) id, &isNew);
+	hPtr = Tcl_CreateHashEntry(&tree->itemHash, (char *)(long) id, &isNew);
 	Tcl_SetHashValue(hPtr, item);
 	tree->itemCount++;
 }
@@ -2015,7 +2014,7 @@ Tree_RemoveItem(
 		Tcl_DeleteHashEntry(hPtr);
 
 	hPtr = Tcl_FindHashEntry(&tree->itemHash,
-			(char *)(intptr_t) TreeItem_GetID(tree, item));
+			(char *)(long) TreeItem_GetID(tree, item));
 	Tcl_DeleteHashEntry(hPtr);
 	tree->itemCount--;
 	if (tree->itemCount == 1)
